==> 1. Prepare PLUTO.ipynb <==
import pandas as pd

### Import data and filter columns

export_cols_1 = ["BBL","BldgClass","LotArea","BldgArea","HistDist","Landmark","OwnerName","XCoord","YCoord"]
export_cols_2 = [col.lower() for col in export_cols_1]

pluto_10 = pd.read_csv('./data/10v2_mn.csv')
pluto_10_c = pluto_10[~pd.isna(pluto_10['BBL'])].copy()
pluto_10_c["BBL"] = pluto_10_c["BBL"].astype("int64")

pluto_10_c[export_cols_1].to_csv('./processed_data/10.csv', index = False)
del pluto_10_c
del pluto_10

pluto_11 = pd.read_csv('./data/11v2_mn.csv')
pluto_11_c = pluto_11[~pd.isna(pluto_11['BBL'])].copy()
pluto_11_c["BBL"] = pluto_11_c["BBL"].astype("int64")

pluto_11_c[export_cols_1].to_csv('./processed_data/11.csv', index = False)
del pluto_11_c
del pluto_11

pluto_12 = pd.read_csv('./data/12v2_mn.csv')
pluto_12_c = pluto_12[~pd.isna(pluto_12['BBL'])].copy()
pluto_12_c["BBL"] = pluto_12_c["BBL"].astype("int64")

pluto_12_c[export_cols_1].to_csv('./processed_data/12.csv', index = False)
del pluto_12_c
del pluto_12

pluto_13 = pd.read_csv('./data/13v2_mn.csv')
pluto_13_c = pluto_13[~pd.isna(pluto_13['BBL'])].copy()
pluto_13_c["BBL"] = pluto_13_c["BBL"].astype("int64")

pluto_13_c[export_cols_1].to_csv('./processed_data/13.csv', index = False)
del pluto_13_c
del pluto_13

pluto_14 = pd.read_csv('./data/14v2_mn.csv')
pluto_14_c = pluto_14[~pd.isna(pluto_14['BBL'])].copy()
pluto_14_c["BBL"] = pluto_14_c["BBL"].astype("int64")

pluto_14_c[export_cols_1].to_csv('./processed_data/14.csv', index = False)
del pluto_14_c
del pluto_14

pluto_15 = pd.read_csv('./data/15v1_mn.csv')
pluto_15_c = pluto_15[~pd.isna(pluto_15['BBL'])].copy()
pluto_15_c["BBL"] = pluto_15_c["BBL"].astype("int64")

pluto_15_c[export_cols_1].to_csv('./processed_data/15.csv', index = False)
del pluto_15_c
del pluto_15

pluto_16_c = pd.read_csv('./data/16v2_mn.csv')
pluto_16_c["BBL"] = pluto_16_c["BBL"].astype("int64")

pluto_16_c[export_cols_1].to_csv('./processed_data/16.csv', index = False)
del pluto_16_c

pluto_17_c = pd.read_csv('./data/17v1_mn.csv')
pluto_17_c["BBL"] = pluto_17_c["BBL"].astype("int64")

pluto_17_c[export_cols_1].to_csv('./processed_data/17.csv', index = False)
del pluto_17_c

pluto_18=pd.read_csv('./data/18v2.csv')
pluto_18_b = pluto_18[pluto_18["borough"] == "MN"]
pluto_18_c = pluto_18_b[~pd.isna(pluto_18_b['bbl'])].copy()
pluto_18_c["bbl"] =pluto_18_c["bbl"].astype("int64")

pluto_18_c[export_cols_2].to_csv('./processed_data/18.csv', index = False)
del pluto_18
del pluto_18_b
del pluto_18_c

pluto_19 = pd.read_csv('./data/19v2.csv')
pluto_19_c = pluto_19[pluto_19["borough"] == "MN"].copy()
pluto_19_c["bbl"] =pluto_19_c["bbl"].astype("int64")

pluto_19_c[export_cols_2].to_csv('./processed_data/19.csv', index = False)

pluto_20 = pd.read_csv('./data/20v1.csv')
pluto_20_c = pluto_20[pluto_20["borough"] == "MN"].copy()
pluto_20_c["bbl"] =pluto_20_c["bbl"].astype("int64")

pluto_20_c[export_cols_2].to_csv('./processed_data/20.csv', index = False)
