==> 2. Combine Years.ipynb <==
import pandas as pd

years = [10,11,12,13,14,15,16,17,18,19,20]

datasets = {}

#format all columns in a standard year, col
for year in years:
    df = pd.read_csv(f'./processed_data/{year}.csv')
    df = df.rename(str.lower, axis='columns')
    
    year_str = str(year)
    renameCols = {}
    for key in ["BldgClass","LotArea","BldgArea","HistDist","Landmark","OwnerName","Xcoord","Ycoord"]:
        renameCols[key.lower()] = f'{year_str}_{key.lower()}'
    df = df.rename(columns = renameCols)
    
    datasets[year] = df

datasets[20].tail()

merged = None
for i, year in enumerate(years):
    if i == 0:
        merged = datasets[year] #init
    else:
        merged = pd.merge(merged, datasets[year], on = "bbl", how = "outer")

merged

#filter for any bbl with a bldgclass that starts with any for any year
years = [10,11,12,13,14,15,16,17,18,19,20]
bldgclass = [f'{year}_bldgclass' for year in years]

m = merged[merged[bldgclass].apply(lambda x: x.str.startswith('M')).any(axis = 1)]

# merged["14"].str.startswith('M') | 
# merged["15"].str.startswith('M') | 

m # 1025

pluto = pd.read_csv('./data/20v1.csv')
mn_pluto = pluto[pluto["borough"] == "MN"][["bbl","cd","council","address","xcoord","ycoord"]].copy()

m2 = pd.merge(m, mn_pluto, on = "bbl", how ="left")

m2.to_csv("./processed_data/filtered_m.csv", index = False)

m2
