==> 3. Fetch DOB data.ipynb <==
import requests
from time import sleep
import pandas as pd
SODA_APP_TOKEN = 'MOkdR2TuLzWqf8P66Yv8CRxZZ'

#import data
m = pd.read_csv('./processed_data/filtered_m.csv')
m.tail()

#fetch DOB Permit Issuance data for every bbl
def getPermits(row):
    #generate url params
    block = str(row['bbl'])[1:6]
    lot = "0" + str(row['bbl'])[6:]

    sleep(0.02)
    r = requests.get(f'https://data.cityofnewyork.us/resource/ipu4-2q9a.json?$$app_token={SODA_APP_TOKEN}&borough=MANHATTAN&block={block}&lot={lot}')
    data = r.json()
    
    #organize data
    permits_array = []
    for permit in data:
        permits_array.append({
            'job_type': permit['job_type'],
            'work_type': permit['work_type'] if 'work_type' in permit else None,
            'date': permit['job_start_date']  if 'job_start_date' in permit else None,
            'filing_status':  permit['filing_status']  if 'filing_status' in permit else None,
        
        })
    print(row.name, row['bbl'], len(permits_array))
    return permits_array

m['dob_permits'] = m.apply(getPermits, axis = 1)

#generate counts for each type (A1, A2, A3, NB, DM, ALL)
def filterJobType(array, job_type):
    return [i for i in array if i['job_type'] == job_type]

def getCounts(row):
    permits = row['dob_permits']
    all_count = len(permits)
    a1_count = len(filterJobType(permits, 'A1'))
    a2_count = len(filterJobType(permits, 'A2'))
    a3_count = len(filterJobType(permits, 'A3'))
    nb_count = len(filterJobType(permits, 'NB'))
    dm_count = len(filterJobType(permits, 'DM'))
    return pd.Series([a1_count, a2_count, a3_count, nb_count, dm_count, all_count])

m[['dob_A1', 'dob_A2', 'dob_A3', 'dob_NB', 'dob_DM', 'dob_ALL']] = m.apply(getCounts, axis = 1)

m[['bbl','dob_A1', 'dob_A2', 'dob_A3', 'dob_NB', 'dob_DM', 'dob_ALL']].tail()

m.to_csv('./processed_data/merged_step3.csv', index = False)
