==> 4. Analysis - PLUTO Build and UnBuilt.ipynb <==
import pandas as pd

#get far fields from lastest pluto
pluto = pd.read_csv('./data/20v1.csv')
mn_pluto = pluto[pluto['borough'] == 'MN'][['bbl', 'builtfar','residfar','commfar','facilfar']].copy()

#merge
df = pd.read_csv('./processed_data/merged_step3.csv')
far_df = pd.merge(df, mn_pluto, on = "bbl", how = "left")

#calculate unbuilt far
def getUnbuilt(row):
    if row['builtfar']:
        max_far = max([row['residfar'], row['commfar'], row['facilfar']])
        return max_far - row['builtfar']
    else:
        return None
    

far_df['unbuiltfar'] = far_df.apply(getUnbuilt, axis = 1)

far_df['unbuiltfar'].tail()

far_df.to_csv('./processed_data/merged_step4.csv', index = False)
