==> 5. Analysis - Adjacent to vacant lots.ipynb <==
Using the latest PLUTO (January 2020 version) there are 904 features classified as Religious Facilities (M1, M2, M3, M4, or M9) in Manhattan. 

1. Using PLUTO’s Shoreline shapefile in QGIS, we selected the Rf’s by filtering the dataset by building class selecting all that beginning with “M” using select by expression.
2. Using select by location to identify ‘vacant lots’ and ‘soft sites’ using the PLUTO’s data dictionary to determine the relevant values to select by. (Appendix) these columns were added to the Rf attribute table.
3. Open field calculator in the RF attribute table was used to identify the that were adjacent to both a vacant lot and a potential soft site.
4. Graphs were created by exporting and using the final Rf attribute utilizing pivot tables.
5. Data shows that there are:
    - 131 RF’s adjacent to vacant lots or potential softsite features.
    - 6 RF’s are adjacent to both a vacant lot and a potential softsite.
    - 58 RF’s adjacent to softsites (total)
    - 79 RF’s adjacent to vacant lots (total)

Appendix
V1-9 Vacant Land
Soft sites include building classes from PLUTO’s DATA Dictionary
Z2 Miscellaneous public parking area,
G0 Residential Tax Class 1 garage
G1 All Parking Garages
G5 Licensed Parking lot
G6 Unlicensed Parking lot

This is consistent with the land use (10) Parking facilities. 
G9- Miscellaneous garage or Gas Station, is not included because it falls under (07)Transportation and utility .


import pandas as pd

#join to processed data 
#to-do convert steps to geopandas
adjacency = pd.read_csv('./processed_data/adjacency.csv')
adjacency[['BBL','Vacant','SoftSite','Adjacency']]

#convert bbl to int
adjacency['BBL'] = adjacency['BBL'].astype("int64")
#lower case all columns
adjacency = adjacency.rename(str.lower, axis='columns')
adjacency = adjacency.drop_duplicates(subset=['bbl'], keep=False)

#join to step4 data
step_4 = pd.read_csv('./processed_data/merged_step4.csv')
step_5 = pd.merge(step_4, adjacency[['bbl','vacant','softsite','adjacency']], on = 'bbl', how="left")

step_5.to_csv('./processed_data/merged_step5.csv', index = False)

len(adjacency)

len(step_4)

len(step_5)
